+++ python/ray/tune/error.py
@@ -7,6 +7,7 @@ class TuneError(Exception):
     """General error class raised by ray.tune."""
     pass
 
+
 class TuneManagerError(TuneError):
     """Error raised in operating the Tune Manager."""
     pass
