+++ python/ray/tune/script_runner.py
@@ -42,7 +42,7 @@ class StatusReporter(object):
         if self._done and not self._latest_result:
             if not self._last_result:
                 raise TuneError("Trial finished without reporting result!")
-            return self._last_result._replace(terminated=True)
+            return self._last_result._replace(done=True)
         with self._lock:
             res = self._latest_result
             self._latest_result = None
